An assignment for a web design class
first servlet allows users to vote on or submit a new type of favorite
music. votes and entries are saved to a glassfish db. user is then
brought to a results page and their vote is saved in the session
variable. the second servlet displays the number of times the user has
voted in that session. this is going from memory so i may have missed
some details. may need to go back and review